backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0.post1
alembic==1.13.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25 